refactor: 클래스명 및 RequestMapping uri 변경
클래스 이름 및 RequestMapping uri 변경
